r-p-s_app/requirements.txt
s2t2-game-utils
python-dotenv
Flask
flask_sqlalchemy
